Make assertion checks work via Marketplace API
I'm not sure how this ever worked!
This bug was maybe revealed as a side effect of
changing the timing of some things (bug 987824).

To be more specific, we always need to send
is_mobile=true because this triggers unverified
assertions. On webpay, assertions are always
unverified as long as it's on mobile.