Fix session to retry failed trans. (bug 1213978)
In case we lose the connection to RabbitMQ/celery (which will raise an exception), we don't want to store the 'already configured' flag in the session.
I'm not 100% sure that sessions are preserved in 500 errors but maybe they are.
If so, this would fix an issue I see where stale transaction IDs are hanging around in people's sessions even after the RabbitMQ issues have been resolved.